Remove assumption that translation tables don't contain capabilities
This was a rather crude way to allow us to ignore memory accesses to
translation tables:  they are exempt from the usual T-CHERI
requirements, and if the translation tables don't contain capabilities,
then we can also ignore them when calculating the set of
available/reachable capabilities.

But we can relax this assumption:  We now instead assume a predicate on
events that indicates whether it is a translation table access or not;
if so, then the T-CHERI requirements are not enforced, but also any
capabilities that it reads are explicitly ignored in available_caps.
Hence, tagged capabilities are allowed in translation tables, but the
non-translation parts of instructions are not allowed to use them
(unless they are also loaded via a regular data load).

The translation event predicate is currently a parameter of the abstract
model, but it might be useful to have read/write kinds for translation
table walks in Sail itself.