Add simpler version of verification helpers
Assumes that assumptions on traces that we want to consider are given as
predicates on individual events that must hold for all events in the
trace, rather than as predicates on full traces.  This should still
allow us to capture many of the assumptions we need (e.g. that the
capability in all register read events from PCC in instruction traces is
unsealed) while disentangling the proof of the T-CHERI properties and
the proof of ISA invariants, potentially simplifying both parts of the
proof.