Further refine privileged register assumptions
If a function or instruction calls a designated function for checking
system register access permission, then it should use the result of that
call to determine whether to access privileged registers or not;  hence,
we do not require that privileged registers are accessible beforehand in
this case.  If that does not work, then it is either a bug, or the
lemmas for that function need to be manually tweaked.