Add bound_untilM, monad_trace_subset and gadgets
The bound_untilM operator is a variant of untilM that has a metric
parameter that must decrease, enforcing termination.

The monad_trace_subset predicate lets us capture a limit (superset)
of the possible elements in the trace. Not very precise, but easy
to prove, and limit e.g. the set of registers touched. Includes a
gadget for automating such proofs.