Add support for a set of initially available capabilities
This normally shouldn't be necessary, as the ISA spec should only use
capabilities that have been directly read from registers or memory, but
allowing an initially available set of capabilities can be useful in
some corner cases.  For example, the Morello ASL spec contains a few cases
of UNKNOWN capabilities that can have arbitrary, nondeterministically
chosen values, only constrained by the prose spec to not increase the
capability rights available to software.  With this patch, we can model
these as capabilities chosen from an arbitrary but fixed subset of the
capabilities reachable from the initial state in the given context.

The monotonicity theorem still holds with the additional assumption that
these initial capabilities are indeed reachable in the initial state.