Non-failure generation with assertion scanning
Add a new exec_success worker predicate which gives us a proof
strategy for non-failure/execution-success that seems to work.

Add an analysis pass to analyse_sail that scans for assert
elements and builds up the expected preconditions of the SAIL
functions. This works and type checks, but hits snags in trying
to export to lem/Isabelle, so maybe it's time to consider another
approach.