Precondition field->param conversion
The idea is to treat a particular field (PSTATE.EL, the exception
level, which is a major despatch point for half the code in
Morello) as always a parameter rather than a dynamic value. This
is working at some high level, though clearly needs tuning.

Also the first of many tweaks to the various proof strategies.