Make capability invocation properties more precise
Better distinguish between the different kinds of capability invocation,
and strengthen requirements, e.g. requiring that capabilities invoked
via indirect sentry are loaded from the address of the sentry plus a
specific offset.  Also require that if an invoked data capability is
written to IDC, then an invoked code capability must be written to PCC.

The monotonicity proof goes through, but the proof tooling for concrete
ISAs still needs fixing.