Tweak idc_write_axiom
Only require that the expected data cap is written if the code cap
written is tagged.  For example, the pair invocation instruction BRS on
Morello might decide to not perform a capability invocation if a check
for Restricted/Executive mode fails, and write an untagged code cap and
sealed data cap instead.